requirements.txt
anyio==3.6.2
arrow==1.2.3
astroid==2.12.13
attrs==22.1.0
beautifulsoup4==4.11.1
binaryornot==0.4.4
certifi==2022.9.24
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
cookiecutter==2.1.1
dill==0.3.6
distlib==0.3.6
exceptiongroup==1.0.4
filelock==3.8.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
identify==2.5.9
idna==3.4
iniconfig==1.1.1
isort==5.10.1
Jinja2==3.1.2
jinja2-time==0.2.0
lazy-object-proxy==1.8.0
MarkupSafe==2.1.1
mccabe==0.7.0
mysql-connector-python==8.0.31
nodeenv==1.7.0
packaging==21.3
Pillow==9.1.1
pkginfo==1.8.3
platformdirs==2.5.4
pluggy==1.0.0
pre-commit==2.20.0
protobuf==3.20.1
pup==1.0.0a17
pylint==2.15.7
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
python-slugify==7.0.0
PyYAML==6.0
requests==2.28.1
requests-mock==1.10.0
requirements-parser==0.5.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
types-setuptools==65.6.0.1
typing_extensions==4.4.0
urllib3==1.26.13
virtualenv==20.17.0
wrapt==1.14.1
zstandard==0.17.0
